Added exception reporting in the templates and cleaned up usage of const
When called with --report-exceptions, kiste2cpp will generate
try-catch-blocks around the called expressions. It will forware the
exception pointers of any caught expcetion to the serializer which can
then do what needs to be done.

If a struct A holds a non-const reference to B, then the B can be
modified in const member functions of A. While this is true in C++, I do
not like particularily and decided to remove the const specifier from
quite a few functions in the serializers and the kiste files.